% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{interaction_plot}
\alias{interaction_plot}
\title{Interaction Effect Plot}
\usage{
interaction_plot(dataset, response, x1, x2)
}
\arguments{
\item{dataset}{the dataset that contains the experiment information}

\item{response}{the value of interest}

\item{x1}{the first factor}

\item{x2}{the second factor}
}
\value{
interaction effect plot
}
\description{
This function gives the interaction effect plots for two treatments in an experiment.
}
