% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{main_plot}
\alias{main_plot}
\title{Main Effect Plots}
\usage{
main_plot(dataset, response, x1, x2, x3, ylim)
}
\arguments{
\item{dataset}{the dataset that contains the experiment information}

\item{response}{the value of interest}

\item{x1}{the first factor}

\item{x2}{the second factor (optional)}

\item{x3}{the third factor (optional)}

\item{ylim}{a vector of the range of y-axis}
}
\value{
main effect plots
}
\description{
This function gives the main effect plots for one or more treatments in an experiment.
If more than 1 treaments are given and ylim is not specified, plots will share the same y-axis range.
}
