% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{var_boxplot}
\alias{var_boxplot}
\title{A box plot to check the equal variances assumption in ANOVA}
\usage{
var_boxplot(dataset, response, x1, x2 = NULL)
}
\arguments{
\item{dataset}{the dataset that contains the experiment information}

\item{response}{the value of interest}

\item{x1}{the first factor}

\item{x2}{the second factor}
}
\value{
a box plot
}
\description{
This function gives a box plots to check the within groups equal variances assumptions.
The first factor x1 is the x-axis, and f the second factor x2 exists,
the box plot will be colored by x2.
}
