% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{var_scatterplot}
\alias{var_scatterplot}
\title{A scatterplot to check the equal variances assumption in ANOVA}
\usage{
var_scatterplot(dataset, response, x1, x2 = NULL)
}
\arguments{
\item{dataset}{the dataset that contains the experiment information}

\item{response}{the value of interest}

\item{x1}{the first factor}

\item{x2}{the second factor (optional)}
}
\value{
a scatterplot
}
\description{
This function gives a scatterplot to check the within groups equal variances assumptions.
The first factor x1 is the x-axis, and if the second factor x2 exists,
the scatterplot will be colored by x2.
}
