% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{var_table}
\alias{var_table}
\title{An interactive table to show variances of different groups}
\usage{
var_table(dataset, response, x1, x2 = NULL, x3 = NULL, type = 1)
}
\arguments{
\item{dataset}{the dataset that contains the experiment information}

\item{response}{the value of interest}

\item{x1}{the first factor}

\item{x2}{the second factor (optional)}

\item{x3}{the third factor (optional)}
}
\value{
an interactive variance table
}
\description{
This function gives an interactive table to show variances of different groups.
It can help check the within groups equal variances assumptions.
This function has two types of table, the default is type 1, and use parameter "type=2" to switch to the second.
In both types of tables, you can click column names to sort.
}
